market-data-service/requirements.txt
# 数据源库
tushare>=1.2.89
akshare>=1.11.0

# Kafka客户端
kafka-python>=2.0.2

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 配置管理
python-dotenv>=1.0.0
pydantic>=2.0.0

# 日志
loguru>=0.7.0

# 调度任务
schedule>=1.2.0

# 异步支持
asyncio-mqtt>=0.13.0

# HTTP客户端
requests>=2.31.0
aiohttp>=3.8.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Flask API
flask>=2.3.0
flask-cors>=4.0.0
pydantic-settings>=2.0.0